adds reducers - caught a few errors
